feat(exercise): add Exercise and ExerciseCatalog components
Add Exercise and ExerciseCatalog components to display a list of exercises
fetched from a server. The Exercise component displays basic information
about an exercise and can be expanded to show additional details. The
ExerciseCatalog component fetches a list of exercises from a server and
displays them using the Exercise component. Also, add styles for the
ExerciseCatalog component.